Move the CSV code to a util header, since it's generally useful.
