+++ src/osvr/Util/CMakeLists.txt
@@ -36,6 +36,7 @@ set(API
     "${HEADER_LOCATION}/ChannelCountC.h"
     "${HEADER_LOCATION}/ClientOpaqueTypesC.h"
     "${HEADER_LOCATION}/ClientReportTypesC.h"
+    "${HEADER_LOCATION}/CSV.h"
     "${HEADER_LOCATION}/DefaultBool.h"
     "${HEADER_LOCATION}/Deletable.h"
     "${HEADER_LOCATION}/DeviceCallbackTypesC.h"
